Add ability to print asset currency
Implemented by overriding __str__() method. This is base for further
work on implementing printing of asset characteristics (for example for
debug mode).

Additionally for pre-committ to work correctly, black code formatter
version had to be increased to 22.3.0 from 22.1.0, because previous
version of black fails to work correctly with click=>8.1.0, see:
https://github.com/psf/black/issues/2976